@@ -6,7 +6,7 @@
   "devDependencies": {
     "babel-cli": "^6.16.0",
     "babel-core": "^6.17.0",
-    "babel-eslint": "^7.1.0",
+    "babel-eslint": "^8.0.1",
     "babel-polyfill": "^6.16.0",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
@@ -16,22 +16,22 @@
     "browserify-css": "^0.9.2",
     "budo": "^9.2.1",
     "cross-env": "^3.1.3",
-    "eslint": "^3.10.1",
-    "react-redux": "^4.4.5",
+    "eslint": "^4.8.0",
+    "react-redux": "^5.0.0",
     "redux": "^3.6.0",
-    "redux-saga": "^0.13.0",
+    "redux-saga": "^0.15.6",
     "rimraf": "^2.5.4",
     "tap-spec": "^4.1.1",
     "tape": "^4.6.2"
   },
   "dependencies": {
     "prop-types": "^15.5.10",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "react": "^15.3.2 || ^16.0.0",
+    "react-dom": "^15.3.2 || ^16.0.0",
     "styled-components": "^1.0.11"
   },
   "peerDependecies": {
-    "redux-saga": "^0.13.0"
+    "redux-saga": "^0.15.6"
   },
   "scripts": {
     "lint": "eslint src",
