@@ -1,5 +1,6 @@
 /*eslint-disable no-unused-vars*/
-import React, { PropTypes } from 'react'
+import React from 'react'
+import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 
 import {
