@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from 'react'
+import React, { Component } from 'react'
+import PropTypes from 'prop-types'
 import CartItem from './CartItem'
 import { connect } from 'react-redux'
 import { checkout, removeFromCart} from '../actions'
